警告つぶし

+++ main.cpp
@@ -2055,7 +2055,7 @@ private:
                 vertex.color = { 1.0f, 1.0f, 1.0f };
 
                 vertices.push_back(vertex);
-                indices.push_back(indices.size());
+                indices.push_back(static_cast<uint32_t>(indices.size()));
             }
         }
     }
